third_party/blink/web_tests/fast/events/constructors/page-transition-event-constructor-expected.txt

This tests the constructor for the PageTransitionEvent DOM class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new PageTransitionEvent('eventType').bubbles is false
PASS new PageTransitionEvent('eventType').cancelable is false
PASS new PageTransitionEvent('eventType').persisted is false
PASS new PageTransitionEvent('eventType', { bubbles: false }).bubbles is false
PASS new PageTransitionEvent('eventType', { bubbles: true }).bubbles is true
PASS new PageTransitionEvent('eventType', { cancelable: false }).cancelable is false
PASS new PageTransitionEvent('eventType', { cancelable: true }).cancelable is true
PASS new PageTransitionEvent('eventType', { persisted: false }).persisted is false
PASS new PageTransitionEvent('eventType', { persisted: true }).persisted is true
PASS new PageTransitionEvent('eventType', { bubbles: true, cancelable: true, persisted: true }).bubbles is true
PASS new PageTransitionEvent('eventType', { bubbles: true, cancelable: true, persisted: true }).cancelable is true
PASS new PageTransitionEvent('eventType', { bubbles: true, cancelable: true, persisted: true }).persisted is true
PASS successfullyParsed is true

TEST COMPLETE
